Move isFixedPitch fix to fix.py
Try saying that three times out loud.
