+++ Lib/gftools/fix.py
@@ -3,6 +3,7 @@ Functions to fix fonts so they conform to the Google Fonts
 specification:
 https://github.com/googlefonts/gf-docs/tree/master/Spec
 """
+from fontTools.misc.fixedTools import otRound
 from fontTools.ttLib import TTFont, newTable, getTableModule
 from fontTools.ttLib.tables import ttProgram
 from fontTools.ttLib.tables._c_m_a_p import CmapSubtable
@@ -49,6 +50,7 @@ __all__ = [
     "drop_mac_cmap",
     "fix_fsselection",
     "fix_pua",
+    "fix_isFixedPitch",
     "fix_font",
     "fix_family",
 ]
@@ -605,6 +607,66 @@ def fix_pua(font):
             ucs4cmap.cmap[0xF0000 + glyphID] = glyph
     font['cmap'] = cmap
     return True
+
+
+def fix_isFixedPitch(ttfont):
+
+    same_width = set()
+    glyph_metrics = ttfont['hmtx'].metrics
+    for character in [chr(c) for c in range(65, 91)]:
+        same_width.add(glyph_metrics[character][0])
+
+    if len(same_width) == 1:
+        if ttfont['post'].isFixedPitch == 1:
+            print("Skipping isFixedPitch is set correctly")
+        else:
+            print("Font is monospace. Updating isFixedPitch to 0")
+            ttfont['post'].isFixedPitch = 1
+
+        familyType = ttfont['OS/2'].panose.bFamilyType
+        if familyType == 2:
+            expected = 9
+        elif familyType == 3 or familyType == 5:
+            expected = 3
+        elif familyType == 0:
+            print("Font is monospace but panose fields seems to be not set."
+                  " Setting values to defaults (FamilyType = 2, Proportion = 9).")
+            ttfont['OS/2'].panose.bFamilyType = 2
+            ttfont['OS/2'].panose.bProportion = 9
+            expected = None
+        else:
+            expected = None
+
+        if expected:
+            if ttfont['OS/2'].panose.bProportion == expected:
+                print("Skipping OS/2.panose.bProportion is set correctly")
+            else:
+                print(("Font is monospace."
+                       " Since OS/2.panose.bFamilyType is {}"
+                       " we're updating OS/2.panose.bProportion"
+                       " to {}").format(familyType, expected))
+                ttfont['OS/2'].panose.bProportion = expected
+
+        widths = [m[0] for m in ttfont['hmtx'].metrics.values() if m[0] > 0]
+        width_max = max(widths)
+        if ttfont['hhea'].advanceWidthMax == width_max:
+            print("Skipping hhea.advanceWidthMax is set correctly")
+        else:
+            print("Font is monospace. Updating hhea.advanceWidthMax to %i" %
+                  width_max)
+            ttfont['hhea'].advanceWidthMax = width_max
+
+        avg_width = otRound(sum(widths) / len(widths))
+        if avg_width == ttfont['OS/2'].xAvgCharWidth:
+            print("Skipping OS/2.xAvgCharWidth is set correctly")
+        else:
+            print("Font is monospace. Updating OS/2.xAvgCharWidth to %i" %
+                  avg_width)
+            ttfont['OS/2'].xAvgCharWidth = avg_width
+    else:
+        ttfont['post'].isFixedPitch = 0
+        ttfont['OS/2'].panose.bProportion = 0
+
 
 def fix_font(font, include_source_fixes=False):
     font["OS/2"].version = 4
